refactor(fixture): 테스트를 위한 fixture 분리
