+++ src/test/java/qna/domain/AnswerRepositoryTest.java
@@ -11,6 +11,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
+import static qna.fixture.TestFixture.JAVAJIGI;
 
 @DataJpaTest
 public class AnswerRepositoryTest {
@@ -27,7 +28,7 @@ public class AnswerRepositoryTest {
 
     @BeforeEach
     void setUp() {
-        user = userRepository.save(new User("kim9418", "123123", "김대겸", "koreatech93@naver.com"));
+        user = userRepository.save(JAVAJIGI);
         question = new Question(1L, "title", "contents");
         question = questionRepository.save(question);
     }
