+++ src/test/java/qna/domain/AnswersTest.java
@@ -4,9 +4,7 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static qna.domain.QuestionRepositoryTest.Q1;
-import static qna.domain.UserTest.JAVAJIGI;
-import static qna.domain.UserTest.SANJIGI;
+import static qna.fixture.TestFixture.*;
 
 class AnswersTest {
