+++ src/test/java/qna/domain/DeleteHistoriesTest.java
@@ -6,8 +6,7 @@ import org.junit.jupiter.api.Test;
 import java.time.LocalDateTime;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
-import static qna.domain.UserTest.JAVAJIGI;
+import static qna.fixture.TestFixture.JAVAJIGI;
 
 class DeleteHistoriesTest {
