+++ src/test/java/qna/domain/DeleteHistoryRepositoryTest.java
@@ -10,6 +10,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import java.time.LocalDateTime;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static qna.fixture.TestFixture.JAVAJIGI;
 
 @DataJpaTest
 public class DeleteHistoryRepositoryTest {
@@ -20,7 +21,7 @@ public class DeleteHistoryRepositoryTest {
 
     @BeforeEach
     void setUp() {
-        user = userRepository.save(new User("kim9418", "123123", "김대겸", "koreatech93@naver.com"));
+        user = userRepository.save(JAVAJIGI);
     }
 
     @Autowired
